refactor: set setup function in one file only
